feat(network) Add ability to exclude accounts in delete default vpc

+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-vpc-stack.ts
@@ -81,7 +81,10 @@ export class NetworkVpcStack extends AcceleratorStack {
     //
     // Delete Default VPCs
     //
-    if (props.networkConfig.defaultVpc?.delete) {
+    if (
+      props.networkConfig.defaultVpc?.delete &&
+      !this.isAccountExcluded(props.networkConfig.defaultVpc.excludeAccounts)
+    ) {
       Logger.info('[network-vpc-stack] Add DeleteDefaultVpc');
       new DeleteDefaultVpc(this, 'DeleteDefaultVpc');
     }
